# Fix performance issues introduced with UserService